% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBsearching.R
\name{DBsearching}
\alias{DBsearching}
\title{Database searching by MSGF+}
\usage{
DBsearching(path, msgfpar, msgfpath)
}
\arguments{
\item{path}{the location containing .raw files and .fasta database}

\item{msgfpar}{specified parameters for MSGF+}

\item{msgfpath}{the location containing MSGF+ java file}
}
\value{
identification information .mzid file
}
\description{
This part contains MSGF+ database searching modules integrated in UHR-IonStar application.
}
