% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceFunc.R
\name{FrameGen}
\alias{FrameGen}
\title{UHR-IonStar Source Functions}
\usage{
FrameGen(db, spectrum, col_filename, col_scannum, accession, pepseq)
}
\description{
These functions are the application functions in the R shiny web app 'UHR-IonStar'.
}
