% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IonStar_DPpipeline.R
\name{IonStar_DPpipeline}
\alias{IonStar_DPpipeline}
\title{IonStar frame-to-protein pipeline}
\usage{
IonStar_DPpipeline(
  raw,
  cond,
  log2form = TRUE,
  norm_method,
  TIC_type = "old",
  frameOR = NULL,
  pepOR = NULL,
  OR_method = "ratio",
  quan_method = "sum"
)
}
\arguments{
\item{raw}{annotated frame file}

\item{cond}{a matrix contains file names and their group information}

\item{log2form}{choose if performing frame normalization on log2 level or not}

\item{norm_method}{choose normalization method ("quantiles" or "TIC")}

\item{quan_method}{choose peptide aggregation method ("sum" or "fit")}
}
\value{
Protein and peptide quantification results
}
\description{
These functions are IonStar frame-to-protein pipeline and its modules
}
