% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newProDataSet.R
\name{newProDataSet}
\alias{newProDataSet}
\title{Format the Frame Dataset}
\usage{
newProDataSet(proData, condition, phenoData = NULL, featureData = NULL)
}
\arguments{
\item{proData}{A frame data set}

\item{condition}{The group information for each file (e.g., c('A','A','A','B','B','B'))}

\item{phenoData}{Former-designed phenotype (default: NULL)}

\item{featureData}{Former-designed feature (default: NULL)}
}
\value{
A 'ExpressionSet' containing frame (peptide) information
}
\description{
This is a former function from package 'IonStarStat'. This function firstly removes frames
that are assigned to different protein groups, then format the dataset to 'ExpressionSet'.
}
