% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierPeptideRM.R
\name{OutlierPeptideRM}
\alias{OutlierPeptideRM}
\title{Outlier Rejection for Peptide Data}
\usage{
OutlierPeptideRM(eset, condition, variance, critM1, critM2, ratio = FALSE)
}
\arguments{
\item{eset}{A ExpressionSet containing peptide abundance information}

\item{condition}{The group information for each file (e.g., c('A','A','A','B','B','B'))}

\item{variance}{Variance}

\item{ratio}{Ratio}
}
\value{
A 'ExpressionSet' containing pepitde information
}
\description{
This is a former function from package 'IonStarStat'.
}
