% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnormalize.R
\name{pnormalize}
\alias{pnormalize}
\title{Normalization for Peptide Data}
\usage{
pnormalize(eset, summarize = FALSE, method)
}
\arguments{
\item{eset}{A ExpressionSet containing peptide abundance information}

\item{summarize}{Summarize}

\item{method}{TIC: total ion current normalization; quantiles: quantile normalization.}
}
\value{
A 'ExpressionSet' containing pepitde information
}
\description{
This is a former function from package 'IonStarStat'. This function performs data normalization at the frame level.
}
