% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteinQuan.R
\name{ProteinQuan}
\alias{ProteinQuan}
\title{Peptide Abundance Aggregating to Protein}
\usage{
ProteinQuan(eset, method = "sum", log2 = TRUE)
}
\arguments{
\item{eset}{A ExpressionSet containing peptide abundance information}

\item{method}{Peptide aggregating method. 'sum': sum abundance; 'fit': general linear mixed model.}

\item{log2}{Whether the data is with log2 form.}
}
\value{
A 'ExpressionSet' containing protein abundance information
}
\description{
These are former functions from package 'IonStarStat' involved in peptide aggregation to protein level.
}
