% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityCheck.R
\name{Quality_check}
\alias{Quality_check}
\title{Quality check}
\usage{
Quality_check(quan, group)
}
\arguments{
\item{quan}{Protein quantification result}

\item{group}{a matrix contains file names and their group information}
}
\description{
This function is for automatically generating quality check plots, including intra-group CV plot and protein intensity rank plot
}
