% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceFunc.R
\name{Ratio_p}
\alias{Ratio_p}
\title{UHR-IonStar Source Functions}
\usage{
Ratio_p(
  quan,
  group,
  decoy,
  control_group = NA,
  test_type = "t",
  equalVar = TRUE
)
}
\description{
These functions are the application functions in the R shiny web app 'UHR-IonStar'.
}
