% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deconvolution.R
\name{shared_checking}
\alias{shared_checking}
\title{Protein Deconvolution Functions}
\usage{
shared_checking(cha, database)
}
\description{
These functions can separate the protein groups to two categories: species-unique and species-shared protein groups
}
