% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedPeptideRM.R
\name{SharedPeptideRM}
\alias{SharedPeptideRM}
\title{Shared Frames Removal}
\usage{
SharedPeptideRM(eset)
}
\arguments{
\item{eset}{A ExpressionSet containing peptide abundance information}
}
\value{
A 'ExpressionSet' containing pepitde information
}
\description{
This are former functions from package 'IonStarStat'. This function removes the frames that are assigned to
multiple protein groups.
}
