% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecRepoGen.R
\name{SpecRepoGen}
\alias{SpecRepoGen}
\title{Spectrum Report Generation}
\usage{
SpecRepoGen(protein, peptide, spectrum, mziddir, MSGFdir, output_name)
}
\arguments{
\item{protein}{target protein file generated by IDpicker.}
}
\value{
A spectrum report
}
\description{
convert mzid files produced by IDpicker to tsv file, then generate spectrum report
}
