% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecRepoGen.R
\name{SpecRepoGen}
\alias{SpecRepoGen}
\title{Spectrum Report Generation}
\usage{
SpecRepoGen(protein, peptide, spectrum, mziddir, MSGFdir, output_name)
}
\arguments{
\item{protein}{target protein file generated by IDpicker.}

\item{peptide}{target peptide file generated by IDpicker.}

\item{spectrum}{target spectrum file generated by IDpicker.}

\item{mziddir}{directory of mzid files}

\item{MSGFdir}{directory of MSGF+ java file}

\item{output_name}{set the name for new PSM list}
}
\value{
A PSM list
}
\description{
Convert mzid files produced by IDpicker to tsv file, then generate a PSM list
}
